--- back-end/pyproject.toml
@@ -1,22 +1,0 @@
-[tool.poetry]
-name = "back-end"
-version = "0.1.0"
-description = ""
-authors = ["Your Name <you@example.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-requests = "^2.32.3"
-pymongo = "^4.7.3"
-google-auth = "^2.30.0"
-PyJWT = "^2.8.0"
-boto3 = "^1.34.129"
-moto = "^5.0.9"
-
-
-[tool.poetry.group.dev.dependencies]
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
