--- back-end/src/planner/http/response.py
@@ -1,12 +1,0 @@
-import json
-
-def response_handler(code, body):
-    return {
-        "statusCode": code,
-        "headers": {
-            "Access-Control-Allow-Headers": "Content-Type",
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Methods": "OPTIONS,POST,GET"
-        },
-        "body": json.dumps(body),
-    }
