+++ backend/pyproject.toml
@@ -1,0 +1,27 @@
+[tool.poetry]
+name = "backend"
+version = "0.1.0"
+description = ""
+authors = ["Your Name <you@example.com>"]
+readme = "README.md"
+package-mode = false
+
+[tool.poetry.dependencies]
+python = "^3.10"
+requests = "^2.32.3"
+pymongo = "^4.7.3"
+google-auth = "^2.30.0"
+PyJWT = "^2.8.0"
+boto3 = "^1.34.129"
+moto = "^5.0.9"
+
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^3.7.1"
+pylint = "^3.2.3"
+black = "^24.4.2"
+pytest = "^8.2.2"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
