+++ front-end/eas.json
@@ -10,4 +10,3 @@
       "production": {}
     }
   }
-
